Use CourseId context, do not guess from series
One Opencast series can be connected to multiple courses.
If we then check the role of the user in that course, we might
discover that the user is not at all in the course that was
guessed from the series resulting in missing permissions
after the LTI login (with no role being set).
Instead of guessing, use the course context ID for embedding
Opencast videos.

GitHub refernce issue:
https://github.com/elan-ev/studip-opencast-plugin/issues/581

Resolves #581