Speed up video sync exist check in discover cronjob (#1239)
We noticed that this cronjob was slow on our test system. The main bottleneck lies in the first video sync task check, as we only migrated a few OC videos and the remaining videos has no preview entry in the database. Therefore, the check is performed on 52994 videos. This patch solves this non-critical problem as it should not exist in the production setup.