Course Default Visibility feature, (#1324)
This PR fixes #1316.
It introduces a new course config via migration 109 called `OPENCAST_COURSE_DEFAULT_EPISODES_VISIBILITY`.
Additionally, a new sidebar action is added which opens a dialog to change the default visibility of episodes in a course. It would apply to the episodes that have "default" as for their visibility!
It works in a way, that it could be set to follow the default "global" config `OPENCAST_HIDE_EPISODES`, visible or hidden!