Refactor migration logic for preview URLs and add fix for orphaned previews,
It considts of 2 parts, first the fix for migration 107, in which the problem was that the update only happened when the new preview has a value, which ends up with not updating those previews that have no search/player.
The second part, is a fix via a new migration step, by which it reads the previews again and try to locate the orphaned previews with json decoded string and then update them either with null or an URL value.