Add metadata change detection to optimize video update process (#1363)
* Add metadata change detection to optimize video update process,

To keep things simple, we first create a non-reactive copy of the event object. We then compare this copy with the current event to detect whether any metadata-related fields have changed. Based on this comparison, we pass a flag to the backend, which decides whether the update metadata call needs to be triggered.

* introducing dirty state of the event & video based on comparison to avoid update metadata call in opencast